pyproject.toml
[tool.poetry]
name = "label_generator"
version = "0.1.0"
description = "Label generator for label printers"
authors = ["Juan Biondi"]
license = "MIT"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "",
  "",
  "",
]

packages = [
    { include = "label_generator" },
]

[tool.poetry.scripts]
label_generator = "label_generator.cli:main_cli"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.1"
Pillow = "^8.3.1"
qrcode = {extras = ["pil"], version = "^7.2"}

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
pytest = "^6.2.4"
pre-commit = "^2.13.0"
bandit = "^1.7.0"


[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | __pycache__
  | \.tox
  | \.venv
  | build
  | dist
  | docs
  | notes
)/
'''

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["label_generator", "tests"]
skip_glob = ["*/setup.py"]
filter_files = true
known_first_party = "label_generator"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
